+++ irohad/ametsuchi/impl/postgres_specific_query_executor.cpp
@@ -35,7 +35,7 @@
 #include "interfaces/queries/get_account_transactions.hpp"
 #include "interfaces/queries/get_asset_info.hpp"
 #include "interfaces/queries/get_block.hpp"
-#include "interfaces/queries/get_engine_response.hpp"
+#include "interfaces/queries/get_engine_receipts.hpp"
 #include "interfaces/queries/get_peers.hpp"
 #include "interfaces/queries/get_pending_transactions.hpp"
 #include "interfaces/queries/get_role_permissions.hpp"
