+++ shared_model/backend/protobuf/CMakeLists.txt
@@ -54,7 +54,7 @@ add_library(shared_model_proto_backend
     queries/impl/proto_asset_pagination_meta.cpp
     queries/impl/proto_account_detail_pagination_meta.cpp
     queries/impl/proto_account_detail_record_id.cpp
-    queries/impl/proto_get_engine_response.cpp
+    queries/impl/proto_get_engine_receipts.cpp
     queries/impl/proto_get_peers.cpp
     queries/impl/proto_ordering.cpp
     )
@@ -79,8 +79,8 @@ if (IROHA_ROOT_PROJECT)
       query_responses/impl/proto_get_block_response.cpp
       query_responses/impl/proto_peers_response.cpp
       transaction_responses/impl/proto_tx_response.cpp
-      query_responses/impl/proto_engine_response_record.cpp
-      query_responses/impl/proto_engine_response.cpp
+      query_responses/impl/proto_engine_receipt.cpp
+      query_responses/impl/proto_engine_receipts_response.cpp
       query_responses/impl/proto_engine_log.cpp
       )
 endif ()
