+++ shared_model/backend/protobuf/query_responses/impl/proto_query_response.cpp
@@ -9,7 +9,7 @@
 #include "backend/protobuf/query_responses/proto_account_detail_response.hpp"
 #include "backend/protobuf/query_responses/proto_account_response.hpp"
 #include "backend/protobuf/query_responses/proto_asset_response.hpp"
-#include "backend/protobuf/query_responses/proto_engine_response.hpp"
+#include "backend/protobuf/query_responses/proto_engine_receipts_response.hpp"
 #include "backend/protobuf/query_responses/proto_error_query_response.hpp"
 #include "backend/protobuf/query_responses/proto_get_block_response.hpp"
 #include "backend/protobuf/query_responses/proto_peers_response.hpp"
