+++ shared_model/backend/protobuf/query_responses/proto_engine_receipt.hpp
