+++ shared_model/backend/protobuf/query_responses/proto_engine_receipts_response.hpp
@@ -8,7 +8,7 @@
 
 #include "interfaces/query_responses/engine_receipts_response.hpp"
 
-#include "backend/protobuf/query_responses/proto_engine_response_record.hpp"
+#include "backend/protobuf/query_responses/proto_engine_receipt.hpp"
 #include "interfaces/common_objects/range_types.hpp"
 #include "qry_responses.pb.h"
