+++ shared_model/interfaces/CMakeLists.txt
@@ -46,7 +46,7 @@ add_library(shared_model_interfaces
     queries/impl/asset_pagination_meta.cpp
     queries/impl/account_detail_pagination_meta.cpp
     queries/impl/account_detail_record_id.cpp
-    queries/impl/get_engine_response.cpp
+    queries/impl/get_engine_receipts.cpp
     queries/impl/get_peers.cpp
     queries/impl/ordering.cpp
     common_objects/impl/amount.cpp
