+++ shared_model/interfaces/queries/impl/get_engine_receipts.cpp
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "interfaces/queries/get_engine_response.hpp"
+#include "interfaces/queries/get_engine_receipts.hpp"
 
 #include "cryptography/hash.hpp"
