+++ test/module/shared_model/query_mocks.hpp
@@ -16,7 +16,7 @@
 #include "interfaces/queries/get_account_transactions.hpp"
 #include "interfaces/queries/get_asset_info.hpp"
 #include "interfaces/queries/get_block.hpp"
-#include "interfaces/queries/get_engine_response.hpp"
+#include "interfaces/queries/get_engine_receipts.hpp"
 #include "interfaces/queries/get_peers.hpp"
 #include "interfaces/queries/get_role_permissions.hpp"
 #include "interfaces/queries/get_roles.hpp"
